requirements.txt
asgiref==3.2.10
authorizenet==1.1.3
Django==3.1.2
django-cors-headers==3.6.0
django-jet==1.0.8
django-treebeard==4.5.1
djangorestframework==3.12.2
mysqlclient==2.0.2
Pillow==8.0.1
psycopg2==2.9.3
python-dateutil==2.8.2
stripe==2.66.0